+++ .gitignore
@@ -15,3 +15,5 @@ coverage/
 lib/
 
 build/
+
+umd/
