+++ packages/icestark-module/webpack.config.js
@@ -1,0 +1,26 @@
+/* eslint-disable @typescript-eslint/no-var-requires */
+const path = require('path');
+
+module.exports = {
+  entry: './src/index.ts',
+  mode: 'development',
+  module: {
+    rules: [
+      {
+        test: /\.tsx?$/,
+        use: 'ts-loader',
+        exclude: /node_modules/,
+      },
+    ],
+  },
+  resolve: {
+    extensions: ['.tsx', '.ts', '.js'],
+  },
+  output: {
+    filename: 'index.js',
+    path: path.resolve(__dirname, 'umd'),
+    library: 'IcestarkModule',
+    libraryTarget: 'umd',
+    libraryExport: "default",
+  },
+};
