+++ packages/icestark-module/webpack.dev.js
@@ -1,0 +1,8 @@
+/* eslint-disable @typescript-eslint/no-var-requires */
+const baseConfig = require('./webpack.config');
+
+module.exports = {
+  ...baseConfig,
+  watch: true,
+  devtool: 'inline-source-map',
+};
